backend/requirements.txt
fastapi
uvicorn
inference-sdk
python-multipart
python-dotenv
